+++ app/Kernel/Data/Flow/FlowIndexData.php
@@ -1,16 +1,14 @@
 <?php
 
-namespace App\Data\Flow;
+namespace App\Kernel\Data\Flow;
 
+use App\Kernel\Data\IndexData;
 use Spatie\LaravelData\Attributes\FromRouteParameter;
-use Spatie\LaravelData\Data;
 
-class FlowShowData extends Data
+class FlowIndexData extends IndexData
 {
     public function __construct(
         #[FromRouteParameter('model')]
         public string $model,
-        #[FromRouteParameter('id')]
-        public string|int $id
     ) {}
 }
