+++ app/Kernel/Data/Flow/FlowShowData.php
@@ -1,14 +1,16 @@
 <?php
 
-namespace App\Data\Flow;
+namespace App\Kernel\Data\Flow;
 
-use App\Kernel\Data\IndexData;
 use Spatie\LaravelData\Attributes\FromRouteParameter;
+use Spatie\LaravelData\Data;
 
-class FlowIndexData extends IndexData
+class FlowShowData extends Data
 {
     public function __construct(
         #[FromRouteParameter('model')]
         public string $model,
+        #[FromRouteParameter('id')]
+        public string|int $id
     ) {}
 }
