+++ app/Kernel/Logics/Flow/FlowLogic.php
@@ -1,12 +1,18 @@
 <?php
 
-namespace App\Kernel\Logics;
+namespace App\Kernel\Logics\Flow;
 
 use App\Kernel\Enums\HttpErrors;
 
 trait FlowLogic
 {
     private string $modelRoute;
+
+    abstract public function allowedModels(): array;
+
+    abstract public function resources(): array;
+
+    abstract public function searchColum(): array;
 
     protected function before(): bool
     {
@@ -15,16 +21,16 @@ trait FlowLogic
         if (! $this->validIsAllowModel()) {
             return false;
         }
-
-        $this->model = new $this->allowedModels[$this->modelRoute];
+        $allowedModels = $this->allowedModels();
+        $this->model = new $allowedModels[$this->modelRoute];
 
         return true;
     }
 
     protected function withResource(): mixed
     {
-        if (array_key_exists($this->modelRoute, $this->resources)) {
-            $resource = $this->resources[$this->modelRoute];
+        if (array_key_exists($this->modelRoute, $this->resources())) {
+            $resource = $this->resources()[$this->modelRoute];
 
             return $resource::collection($this->response);
         }
@@ -34,7 +40,7 @@ trait FlowLogic
 
     public function validIsAllowModel(): bool
     {
-        if (array_key_exists($this->modelRoute, $this->allowedModels)) {
+        if (array_key_exists($this->modelRoute, $this->allowedModels())) {
             return true;
         }
