+++ app/Logics/Flow/FlowIndexLogic.php
@@ -2,12 +2,21 @@
 
 namespace App\Logics\Flow;
 
+use App\Kernel\Data\Flow\FlowIndexData;
 use App\Kernel\Logics\Flow\FlowIndexLogicBase;
 use App\Models\Action;
 use App\Models\Role;
+use Illuminate\Http\JsonResponse;
+use Spatie\LaravelData\Data;
 
 class FlowIndexLogic extends FlowIndexLogicBase
 {
+    protected Data|FlowIndexData $input;
+
+    public function run(Data|FlowIndexData $input): JsonResponse
+    {
+        return parent::logic($input);
+    }
 
     public function allowedModels(): array
     {
